create recipe page, index link and assets folder
The recipe grilled chicken breast is the first recipe added into the
website. It can be accessed through relative link in index.html. It
will serve as a template for other recipes that might be added in the
future. Assets folder is created to store assets such as images of the
food according to the recipe.